LPS-167515 Move logic to display context (aliases)
